Switch to standalone conda installer on AWS Batch
Conda recently started shipping a stand-alone installer.
It is smaller than the current binary that we pull in (~10mb vs ~60mb)
as well as saves time since we don't have to install conda by itself.

As a follow-up, we can even dump the installer in the user's S3 bucket
to save this network ingress on AWS Batch.