Set namespace to None in example notebook
Setting namespace to None. A common scenario is where the user
copies to contents of the notebook into a sagemaker notebook and executes
the code snippets. That would fail, because the sagemaker notebook runs
with user:ec2-user. Setting the namespace to None allows us to bypass this
issue and provide a better user experience.