Improve IncludeFile support to support Step-Functions (#220)
* Improve parameters by adding multiple defaults (depending on the top level CLI command)
as well as having an explicit load_parameter function

This is to support IncludeFile for Step Functions.

* [WIP] Make IncludeFile compatible with Step Functions

This also allows IncludeFile to directly include a S3 file. Note
that the S3 file is not persisted separately so it will be accessed
everytime a run is launched. In other words, you should treat it like
an external table/database which may change between runs.

* Add a new "Local" datatool to read/write to the local filesystem

This mirrors the S3 datatool. This change is to support a future IncludeFile
semantic which will rely on datatools to load/store files.

* Improve IncludeFile to re-enable local files.

When specifying an include file, you can now use the following syntax:
  - scheme://<path> (for eg: s3://<path> or local://<path>)
  - an absolute or relative local file

In the first case, the file is *not* persisted to the datastore. It can effectively be viewed
as an external separately managed artifact. If used in your run, it will be loaded at that time (so
will have whatever value happens to be there when the task uses the IncludeFile). The file is therefore
materialized on use. It is *never* stored in the datastore unless you explicitly re-assign it
to self:
  @step
  def my_step(self):
    self.save_file = self.my_file

In the second case, the local file is persisted to the datastore that you are currently using.
For example, if you have a S3 datastore, the local file will be uploaded to S3. If you have
a local datastore, the local file will be copied to the local datastore. This effectively takes
a snapshot of the file. The timing of the snapshot is as follows:
  - on run, the file is snapshoted prior to the start step
  - on step-functions create, the default file is snapshoted when the step function is deployed
  - on step-functions trigger, the file passed as an argument is snapshoted when triggering

* Simplify local file handling by integrating with IncludeFile for now; revert to deploy_mode

Also contains a few other fixes to address comments

* Fix deploy-time functions value for --help

* Some changes

* Uncomment change

Co-authored-by: Romain Cledat <rcledat@netflix.com>
Co-authored-by: Savin <savingoyal@gmail.com>