Guard against inconsistently cased filesystems for @conda on macOS with @batch
Conda fails to correctly set up environments for linux-64 packages on macOS which
is needed to collect the necessary metadata for correctly setting up the conda
environment on AWS Batch. This patch simply ignores the error-checks that conda
throws while setting up the environments on macOS when the intended destination
is AWS Batch.