Support very wide workflows in AWS Step Functions (#403)
Very wide workflows - `split-and` might cause issues with
AWS Step Functions because of the limitation imposed on step
name size. Changing the step name for a `Parallel` task type
in AWS Step Functions to a hashed value buys us a lot of
flexibility. This is not a user-visible change since the Parallel
step is an implementation detail of the Metaflow -> SFN compilation
process