Support multi-level foreach
In order to design the nested foreach in Argo, we need to use nested DAG.

Each node in metaflow that has these two characteristics, counts as a nested foreach.
First, the is_inside_foreach for this node is True.
Second, its parent has type foreach.
Therefore, this node uses as the entry node for nested DAG.
Moreover, we need to specify the exit condition(node) from this nested DAG.
The exit condition is when we reach to the matching join of the entry node's parent.
But before building nested dag, we need to do topological sort on graph and passed
the ordered queue to build the nested DAG template. The reason for sorting is if both nested foreach
and nested branch happens in this nested foreach. We should be sure that the branches and their
parent must be add in the same DAG. Consider is_inside_foreach for node Y is True and its parent X
has a foreach type. First, we sort the steps from Y to maching_join of X (join_x). Then, define the
nested template which includes inputs,outputs and DAG and add Y as the entry node for this DAG and
iterate over ordered queue till we reach the join_x. Here, the iteration stops and return the output
of the latest node, which is added to this nested DAG,to the upper DAG. Because in Argo the nested
DAG must pass the output to the upper DAG (the DAG that this nested DAG calls inside it).
The next node after each nested dag is always a join node (exit_node). This node doesn't follow the
default rule for input-path and dependencies. Its input-path as follows:
'{{workflow.name}}/(join.in_funcs)>/{{tasks.<(split_parent(join).out_func)>.outputs.parameters}}'
and its dependencies is split_parent(join).out_func.