Introduce template for UnboundedForeach.
* As denoted by plugins/test_unbounded_foreach/* this outlines a
template for creating (unbounded) foreach which requires a control
task that baby-sits the execution of a foreach while delegating the
actual execution of the foreach task possibly to a separate service.

* Concrete use-cases include a hyperparameter tuning service, or a
different service implementing a map-reduce paradigm but with a
specialized purpose.

TODO:
* Enable tests in local mode after bugs in datastore (py2, py3 incompat);
local metadata service (w.r.t task registration) are fixed.
* CliChecker might have to be updated to allow accessing a run due to
the presence of local metadata provider.