Handle hyphens in parameter names for AWS Step Functions executions
Prior to this PR, an invocation like
`python flow.py step-functions create --foo-bar 42` will result in
an error. Metaflow behind the scenes relies on environment variables
to correctly pass parameters to various steps. The environment
variables unfortunately can't have hyphens in the name - this PR
acconts for that behavior and fixes the issue.