Add check for empty describe_jobs response for AWS Batch
It seems that the response for describe_jobs could be empty,
particularly when the job has been just launched. Although, I
haven't been able to replicate the issue on my end, a handful of
Metaflow users seem to have run into it. When such an issue happens,
this PR happily ignores the response and tries again.