Cache the S3 boto connection in the datastore (#570)
* The datastore now caches the S3 boto connection and passes it to the S3 client

This allows us to use S3() using with (and therefore properly clean up the temp
directory) but also benefit from the connection reuse.

As part of this change, we also had to change the interface to load_bytes. This
change is invisible to the user as it is 100% internal to the datastore.

Cleaned up the s3tail.py and s3util.py files to co-locate them with the
datatools instead of the datastore (since they were no longer used directly
in the datastore). This simplified the import story.

* Addressed comments and simplified the code a bit

- Use a single CloseAfterUse class.
- Remove the use of caching on write (in artifact and blob cache); the cache
  now only caches things that are read which is the way it is being used
- Simplified the types of caches used in datastore_set.py

* Remove no longer needed file

* Clean up usage of CloseAfterUse

* remove artifact_cache, simplify blob_cache

* get rid of redundant LazyFile

* fix FileBlobCache root

* fix paths in content_addressed_store

* Addressed comments

* Move check on file existence in CAS to backend for better efficiency

* Fix modes for datastore due to merge_artifacts functionality

* Ignore error when overwrite is False and file exists

Co-authored-by: Ville Tuulos <tuulos@gmail.com>