Tweaks and optimizations to datastore
* Revert "Move check on file existence in CAS to backend for better efficiency"

This reverts commit d0ac6fe228217b7429893f634eb7e07793aabe9b.

* Change encoding type to gzip+pickle to keep forward compatibility

* Reduce memory usage of artifact save and load

We ensure that we do not keep around temporaries and allow the
GC to remove objects that are no longer needed

* Make is_file parallel

We can now call is_file in parallel. It is unclear if this helps a lot given the
overhead of loading the s3op and starting them all. It may help in the case of
a very large number of artifacts

* Set overwrite=True when uploading metadata

* Improve fetching info and file at the same time

* reduce duplicate copies in datastore by using generators

* minimize memory consumption by handling artifacts one by one

* fix docstrings and documentation

* fix @batch to work with the new datastore

* do not implicitly assume a correct key order in flow_datastore.load_data

* Other minor consistency tweaks

Co-authored-by: Romain Cledat <rcledat@netflix.com>