Address another issue with load_artifacts
When an artifact is present in the blob cache, load_blobs would return
things out of order and load_artifacts would therefore return the wrong
blob for the artifact name.

This patch hopefully addresses this issue fully and also adds a small
optimization whereas blobs are not requested multiple times if they are
identical.