Better handling of initial context message for sidecars
In the previous implementation, a failure to send the initial context
message could crash the flow. This change properly handles that message
like any other message (so has no impact on the flow itself). It also
implements better retry policies to send the context. Finally, this change
also improves error handling: if a message fails to send, in the previous
implementation, it was likely that the next successful one would be
an invalid json.