Change behavior of env escape when module is present
When a module configured in the env escape was present in the local
environment, we would raise an error. We now print a message and use
the module present in the environment (not the environment escape).
This allows the env escape modules to provide a "fallback" in case the
module is not present but to still use the local module if present