Fix an issue with the environment escape server directory
The environment escape server would launch in the cwd. This could
cause issues if, for example, a `metaflow` directory existed in that
cwd (because it would then take precedence over the actual metaflow.

To remediate this situation, the server is launched in the configuration
directory (where we know what there is).

Also took care of a quick annoyance of having to create an empty overrides.py
file for everything. This is now no longer required