Clean up what values are passed down
Now, if a value is the default, it is not propagated by default since it
can be reconstructed simply from the code. With this change, we basically
propagate (by default) only those values that require some external knowledge
(env vars, configuration file) to set. This reduces the set of values
propagated to only the ones the user overrides.