Clean up configuration variables (#1183)
* Clean up configuration variables

* Fix bug in argo/airflow; update default value of AZURE_STORAGE_WORKLOAD_TYPE

* Forgot file

* Clean up what values are passed down

Now, if a value is the default, it is not propagated by default since it
can be reconstructed simply from the code. With this change, we basically
propagate (by default) only those values that require some external knowledge
(env vars, configuration file) to set. This reduces the set of values
propagated to only the ones the user overrides.

* Remove propagate flag; change name of SERVICE_URL for batch

* Fixed more instances of INTERNAL_SERVICE_URL; fixed comments