Deal with transient errors (like SlowDowns) more effectively for S3 (#1186)
* Deal with transient errors (like SlowDowns) more effectively for S3

In the previous incantantion, SlowDown errors were treated as a regular
error and everything was retried (making things worse). With this change,
we continue retrying the operations that were unsuccessful (and only those).

Also use a better internal boto retry policy (unless specified by the user).

Modified the tests to inject failures to be able to test the functionality. Tests
now have a failure injection rate of 0, 10, 50 or 90 percent.

* Added more comments

* Addressed comments; added more comments in code