Fix an issue with threading and the escape hatch.
The escape hatch's mechanism to communicate with the "escaped" python
process is not thread safe. This addresses this issue by properly
locking when communicating with the backend process.

A more long term fix may be to also multithread the backend process
and properly multiplex communication but it is unclear what the actual
benefit would be as this mechanism is more frequently used in a single
threaded manner.