When packaging `metaflow_extensions`, add an empty `__init__.py` file.
This will force the packaged `metaflow_extensions` package to not
be a namespace package and therefore prevent python from looking for
other extensions that may be present on the remote system thereby
giving the user the exact same metaflow environment as the one they
had locally.

In particular, this closes off a sort of "code injection" possibility.