Fix the `logs` command in cases where the step/task hasn't finished (#1315)
* Fix the `logs` command in cases where the step/task hasn't finished

When the `logs` command was invoked, it would go and figure out the
task_datastores it needed to get the logs from. In some cases (when
an attempt was started but not finished), it would try to access
the data file for the datastore (even though it didn't exist). It is
actually not needed in this particular case so this PR modifies the code
to ignore that.

* Doh