Fix #1326 and strips ext_info from blobs passed to schedulers
The `ext_info` field is only needed in the INFO file for extension
support and is not gleaned from the environment passed to the schedulers.
We also remove it from the tags so it won't be stored in the DB (also of limited
value there -- the extensions installed are already present in the version
string).