Reduce multiple metadata service rtts using cached version.
The code for starting the heartbeat calls version() multiple
times even when the actual value of the version is already available.
Each call to version() makes a round-trip to the metadata-service
which increases the overall latency.

This change simply uses the value of the first call to version() in
subsequent checks avoids 2 extra rtts the metadata service.