Job completion check based on container status.
This change modifies the way a Metaflow step running as a Kubernetes job
is marked completed. Previously, Metaflow would wait until the job
itself is marked as completed. However, this can take much longer than
the time the container running in the job is marked completed.

There don't seem to be any obvious tunables in K8s for making it faster
to mark a job as completed.

Given that a Metaflow job runs a single container, this change relies
on checking that containers' status.

Testing Done:

- Basic linear flows run to completion.
- Foreach based flows run to completion.
- Flows that have failed steps show up as failed. Logs were shown on
  stderr and the metaflow ui.
  - Tested with a step that explicitly did a sys.exit()
  - Tested with a step that had it's pod explicitly terminated
    (kubectl delete pod)
- Terminated one step in a foreach. Verified that the run
  shows up as failed
- Tested that flows run just fine with Argo workflows