Fix an issue when LD_LIBRARY_PATH is overriden in a Conda environment
In a Conda environment, it is sometimes necessary to set LD_LIBRARY_PATH to
first include the Conda's environment libraries before anything else.

This, however, causes problems with the escape hatch which uses this same
modified LD_LIBRARY_PATH *outside* of the Conda environment. This patch
fixes this issue by using the convention that, if a Conda environment
modifies LD_LIBRARY_PATH, it will save the original LD_LIBRARY_PATH
in MF_ORIG_LD_LIBRARY_PATH.