Fix two possible issues with the environment escape communication (#1555)
The first issue is when subprocesses are created after the main process has
activated the environment escape. In this case, multiple processes could try
to communicate wit the single escaped server which is a big no no. We explicitly
block this from happening now. The case we ran into was an object allocated in a parent
being gc in a subprocess.

The second issue relates to the GC which could trigger *while* another communication
was happening. This could result in a deadlock since the lock protecting the communication
is not reentrant (on purpose). We solve this by preventing the GC from being active during
that time.