[argo] Add k8s labels to tempalte.metadata instead of spec.podMetadata
Argo does not propagate the spec.podLabels to the pods.
Only the sections under templates from the WorkflowTemplate are
propagated to the Workflow and the pods. Their docs are misleading.

Here are related argo issues,
1. https://github.com/argoproj/argo-workflows/issues/7091
2. https://github.com/argoproj/argo-workflows/issues/7013

Particularly, this [comment](https://github.com/argoproj/argo-workflows/issues/7091#issuecomment-978687939)
is helpful in demystifying some confusion of the argo terminology

Signed-off-by: jhasm <santosh.m.jha@gmail.com>