Get OOM errors to stderr and the UI
There is a race between Metaflow detecting that a pod failed execution
and the reason for pod failure being set on the pod. As a result,
at times, the failure reason doesn't posted to stderr.

This change makes Metaflow try a little harder to get the reason for
the failures.

For pods that get OOM killed, this change worked just fine.

Without this change, the OOM killed pod would simply die and the user
would have no idea why. With this change, the error on stderr shows:

Task ran out of memory. Increase the available memory by specifying @resource(memory=...) for the step.