Sanitize path names when creating cache directories in FileCache (#1931)
We didn't remove the `/` from names and it seems to be creating issues in
some cases.

Note that this would no longer create a deep hierarchy. This may cause issues
for very many many flows. We could keep the "/" and just strip the schema
but am not sure if the issue would fully go away.