Add installed_extensions information (#2018)
* Add installed_extensions information

This adds information to the system tags about the metaflow extensions that
are installed. This makes it easier to reproduce the exact same
environment. Information about whether the package information is complete or
not is also included.

This PR also fixes some issues with extension loading and allows extension
information to be gathered in a programatic fashion thereby enabling the
discovery of extensions.

* Allow users to configure where extensions are searched

* Move EXTENSIONS_SEARCH_DIRS out of config

* Move the extension installation information to _graph_info