Modify getting Metaflow version (#2009)
* Modify getting Metaflow version

Favors reading the INFO file if present to be able to have the
most accurate version of Metaflow when executing remotely (especially in
the presence of extensions).

Also limit reading the INFO file to once per process (as opposed to possibly
twice).

Finally, gets the version of the source of Metaflow (and not the current
directory)

* Addressed comments

* Add support for overriding version string

* Fix issue with brew (hopefully)

* Revert "Add support for overriding version string"

This reverts commit 0434e975d0ffc20c601a460e8a33c49e663214ac.

* Add installed_extensions information (#2018)

* Add installed_extensions information

This adds information to the system tags about the metaflow extensions that
are installed. This makes it easier to reproduce the exact same
environment. Information about whether the package information is complete or
not is also included.

This PR also fixes some issues with extension loading and allows extension
information to be gathered in a programatic fashion thereby enabling the
discovery of extensions.

* Allow users to configure where extensions are searched

* Move EXTENSIONS_SEARCH_DIRS out of config

* Move the extension installation information to _graph_info

* Use .metadata[X] for name and version since not all importlibs have the aliases