[kubernetes][parallel] hostname resolution timeout
In Kubernetes, jobsets ensure that all jobs within the set are scheduled together. However, while they may be scheduled concurrently, the order in which they start executing is not guaranteed. This can cause issues if a worker job starts before the control job, as the worker may fail to resolve the control job’s hostname, leading to a crash.

This update introduces a configurable `hostname_resolution_timeout`, allowing worker jobs to wait for a defined period to resolve the control job's hostname before terminating. This adds a layer of resilience, preventing crashes caused by timing issues between job start times.

One thing to note here is that if the control job finishes before the worker job starts, the worker will still wait for the `hostname_resolution_timeout`. Given that many higher-level decorators using @parallel already implement distributed barriers, this change ensures that the execution order does not jeopardize the success of the jobset.