Clean up code packaging
Separating out the Tar backend portion so it can be replaced
by something else. Internally, at Netflix, we also have the need for
a zip packaging and separating it out like this make it possible.

The code in MetaflowPackaging is also much cleaner and clearer focusing
simply on *what* to package and not specifically on the how (moving out
all the walking code, the actual taring, etc).