Fixes tolerations typehint (#2502)
This is a very minor fix - but it confused me for a while, so for the
next person poking at the code, this might be a good to sort out. The
format of the tolerations is a dict with key and value strings. It is
listed correctly in a comment further down in the codebase but
incorrectly in the typehint.