feature: conditionals for Argo Workflows (#2550)
Adds Conditional support to Argo Workflows

TODO:

- [x] nesting a conditional as the last step inside a foreach is not yet
working. This is now out of scope for the initial release
- [x] test `@parallel` with conditionals
- [x] test regular flows with argo for possible regressions

Open issues to solve in another PR:
- conditionals as last step in a foreach. requires thorough input-paths
rework
- `@parallel` as a switch step, where it leads to more than one possible
parallel-join step. This needs to either be fixed, or blocked. Issue
exists with local conditional runs as well so outside of the scope for
this PR