Fix extensions and packaging for some types of editable package installs
In some cases, a distribution install of an editable package includes just
a .pth file and not the full file list included in the extension. In those
cases, the extension mechanism failed to fall back to traversing the
directory indicated in the .pth file.

This PR fixes this issue while trying to maintain the "usual"
case of non editable packages as quick as possible (walking the FS
is not ideal in all cases).

Some additional clean-up particularly around handling the case of
"getting file from the distribution files list" and "walking a tree". We
already used to walk the FS for non-distribution packages but now we do
it in both cases (potentially) so the code was refactored to make it
clearer.

A similar issue could also happen when packaging METAFLOW_PACKAGE_POLICY
packages that were installed in an editable mode. The same type of fixes
are applied there as well.